Abstracted away the need for the base BinaryLogConsumer trait to know about host, port, user, password.
This information is only needed for the MySQL binary log consumer and
was spilling out and breaking encapsulation in the Conf and Pipe areas
of the code (having to poke into artificially public members). We now
have a consumer ID that is enforced at the BinaryLogConsumer level that
will be used to replace passing host/port around.