Added support to conditionally fail upon encountering a mutationless transaction.
By default, the behaviour can be controlled by the configuration flag
quit-on-empty-mutation-commit-failure which is set to false. This can be
overriden with custom error handling that has access to all the unknown
queries that have been accumulated in the transaction.

Quickly fixed TableCacheSpec since it was failing to test for table
refreshes.