Better handling of primary key values in tables.
Previously, if a table has no primary keys we would end up with a Table
object with a primaryKey=Some(List()) which is not desirable since the code
expects there to be a None if the table does not contain a declared
primary key.

The standard out producer does not treat updates and deletes to tables
without primary keys properly. If a table with 3 rows is updated, we
will have something along the following lines on standard out:

    UPDATE mypipe.foo SET (id=3)
    UPDATE mypipe.foo SET (id=3)
    UPDATE mypipe.foo SET (id=3)