Added ability to only process specific dbs / tables. #11
The configuration now has a setting that controls what dbs and tables to consume.

Effectively, this is treated as Scala code and is compiled at run time. Both "db" and "table" are String types. If this value is set to blank then it will be ignored (and all dbs / tables will be processed).

Example (triple quote on purpose for inner quotes):

    """ db == "mypipe" && table =="user" """