Added ability to supply "--no-transaction" to snapshotter. #28
This performs the snapshot without a consistent read and without a
transaction.

The idea here is as follows (credit goes to @FelixGV):

- Record the DB's current bin log position.
- Do a full scan of the table(s), producing each row into your Kafka
topic.
- Replay the bin log from the position recorded in step 1.
- Optionally, record the Kafka offsets when you reach the tip of the bin
log (or when you are within a certain lag threshold, say <5 seconds from
the current time or whatever).
- Let Kafka log compaction take care of eliminating duplicates keys.

Once compaction runs once, the topic's content would be fully ordered.
Before log compaction runs (or if you don't want to run it at all), then
you need to make sure you consume at least up to the Kafka offset
recorded in step 4, because the events prior to that are potentially out
of order.