Removed jersey-server dep (not needed), added application.example.conf.
Moved and made better the content of applicaiton.conf into
application.example.conf simply leaving an include statement in there.
This is so that the configuration does not get merged in case an
override is used resulting in potentially useless and failing config
entries.

Also changed mypipe-kafka's logback test log level to info from debug.